\name{smoother}
\alias{smoother}
\title{Smooth Data in R}
\description{
\code{smoother} package for the smoothing of numerical
data.
}
\details{
\code{smoother} is presently limited to a port of the
Matlab 'Gaussian Window' Function, which can be found at
the following URL: \url{http://goo.gl/HGM47U} and should be
used on an 'as-is' basis.

For further information (and examples) with regards to
utilizing the gaussian window function, please refer to the
\link{smth} function help file
}
\references{
The Gaussian Smoothing component of the \code{smoother}
package has been loosley adapted from the following works:
\url{http://goo.gl/NK79bJ}. In this work, the two tail
regions of the data are handled a little differently.
}

